vehicle-spec-rag/requirements.txt
pymupdf
tqdm
spacy
pandas
sentence-transformers
chromadb
google-generativeai
python-dotenv
fastapi
uvicorn
